fix: move complex code in a file of its own
